+++ lib/reduced.dart
@@ -3,7 +3,7 @@
 export 'src/event.dart';
 export 'src/store.dart';
 export 'src/callable.dart';
-export 'src/carrier.dart';
+export 'src/action.dart';
 export 'src/parametrized.dart';
 export 'src/function.dart';
 export 'src/creator.dart';