+++ test/carrier/action3_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:reduced/src/carrier.dart';
+import 'package:reduced/src/action.dart';
 import 'package:reduced/src/event.dart';
 import 'package:reduced/src/store.dart';
 
@@ -7,7 +7,8 @@ class MockEvent3 extends Event3<String, String, String, String> {
   MockEvent3();
 
   @override
-  call(state, value1, value2, value3) => '$state $value1 $value2 $value3';
+  call(state, value1, value2, value3) =>
+      '$state $value1 $value2 $value3';
 }
 
 class MockStore extends Store<String> {
@@ -21,9 +22,9 @@ class MockStore extends Store<String> {
 }
 
 void main() {
-  test('Event3Carrier call test', () {
+  test('Action3 call test', () {
     final store = MockStore('0');
-    final objectUnterTest = Event3Carrier(
+    final objectUnterTest = Action3(
       store,
       MockEvent3(),
     );
@@ -31,50 +32,53 @@ void main() {
     expect(store.state, '0 1 2 3');
   });
 
-  test('Event3Carrier hashCode test', () {
+  test('Action3 hashCode test', () {
     final store1 = MockStore('1');
     final reducer1 = MockEvent3();
     final store2 = MockStore('2');
     final reducer2 = MockEvent3();
-    final objectUnterTest11 = Event3Carrier(
+    final objectUnterTest11 = Action3(
       store1,
       reducer1,
     );
-    final objectUnterTest12 = Event3Carrier(
+    final objectUnterTest12 = Action3(
       store1,
       reducer2,
     );
-    final objectUnterTest21 = Event3Carrier(
+    final objectUnterTest21 = Action3(
       store2,
       reducer1,
     );
-    final objectUnterTest22 = Event3Carrier(
+    final objectUnterTest22 = Action3(
       store1,
       reducer1,
     );
     expect(objectUnterTest11.hashCode, objectUnterTest22.hashCode);
-    expect(objectUnterTest11.hashCode, isNot(objectUnterTest12.hashCode));
-    expect(objectUnterTest11.hashCode, isNot(objectUnterTest21.hashCode));
-    expect(objectUnterTest12.hashCode, isNot(objectUnterTest21.hashCode));
+    expect(objectUnterTest11.hashCode,
+        isNot(objectUnterTest12.hashCode));
+    expect(objectUnterTest11.hashCode,
+        isNot(objectUnterTest21.hashCode));
+    expect(objectUnterTest12.hashCode,
+        isNot(objectUnterTest21.hashCode));
   });
-  test('Event3Carrier operator== test', () {
+  test('Action3 operator== test', () {
     final store1 = MockStore('1');
     final reducer1 = MockEvent3();
     final store2 = MockStore('2');
     final reducer2 = MockEvent3();
-    final objectUnterTest11 = Event3Carrier(
+    final objectUnterTest11 = Action3(
       store1,
       reducer1,
     );
-    final objectUnterTest12 = Event3Carrier(
+    final objectUnterTest12 = Action3(
       store1,
       reducer2,
     );
-    final objectUnterTest21 = Event3Carrier(
+    final objectUnterTest21 = Action3(
       store2,
       reducer1,
     );
-    final objectUnterTest22 = Event3Carrier(
+    final objectUnterTest22 = Action3(
       store1,
       reducer1,
     );