+++ test/carrier/action_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:reduced/src/carrier.dart';
+import 'package:reduced/src/action.dart';
 import 'package:reduced/src/event.dart';
 import 'package:reduced/src/store.dart';
 
@@ -26,7 +26,7 @@ class MockStore extends Store<Object> {
 void main() {
   test('EventCarrier call test', () {
     final store = MockStore(0);
-    final objectUnterTest = EventCarrier(
+    final objectUnterTest = Action(
       store,
       MockEvent(1),
     );
@@ -34,50 +34,53 @@ void main() {
     expect(store.state, 1);
   });
 
-  test('EventCarrier hashCode test', () {
+  test('Action hashCode test', () {
     final store1 = MockStore(1);
     final reducer1 = MockEvent(1);
     final store2 = MockStore(2);
     final reducer2 = MockEvent(2);
-    final objectUnterTest11 = EventCarrier(
+    final objectUnterTest11 = Action(
       store1,
       reducer1,
     );
-    final objectUnterTest12 = EventCarrier(
+    final objectUnterTest12 = Action(
       store1,
       reducer2,
     );
-    final objectUnterTest21 = EventCarrier(
+    final objectUnterTest21 = Action(
       store2,
       reducer1,
     );
-    final objectUnterTest22 = EventCarrier(
+    final objectUnterTest22 = Action(
       store1,
       reducer1,
     );
     expect(objectUnterTest11.hashCode, objectUnterTest22.hashCode);
-    expect(objectUnterTest11.hashCode, isNot(objectUnterTest12.hashCode));
-    expect(objectUnterTest11.hashCode, isNot(objectUnterTest21.hashCode));
-    expect(objectUnterTest12.hashCode, isNot(objectUnterTest21.hashCode));
+    expect(objectUnterTest11.hashCode,
+        isNot(objectUnterTest12.hashCode));
+    expect(objectUnterTest11.hashCode,
+        isNot(objectUnterTest21.hashCode));
+    expect(objectUnterTest12.hashCode,
+        isNot(objectUnterTest21.hashCode));
   });
-  test('EventCarrier operator== test', () {
+  test('Action operator== test', () {
     final store1 = MockStore(1);
     final reducer1 = MockEvent(1);
     final store2 = MockStore(2);
     final reducer2 = MockEvent(2);
-    final objectUnterTest11 = EventCarrier(
+    final objectUnterTest11 = Action(
       store1,
       reducer1,
     );
-    final objectUnterTest12 = EventCarrier(
+    final objectUnterTest12 = Action(
       store1,
       reducer2,
     );
-    final objectUnterTest21 = EventCarrier(
+    final objectUnterTest21 = Action(
       store2,
       reducer1,
     );
-    final objectUnterTest22 = EventCarrier(
+    final objectUnterTest22 = Action(
       store1,
       reducer1,
     );